\name{boxTidwell}
\title{Box Tidwell transformation...}
\usage{'\%bt\%'(x, pow)}
\description{Box Tidwell transformation}
\details{Simple function to apply the Box Tidwell transformation to a vector, when a (single) power is
provided.}
\value{the transformed vector}
\alias{boxTidwell}
\alias{'\%bt\%'}
\keyword{utilities}
\keyword{internal}
\arguments{\item{x}{the numeric vector}
\item{pow}{the single power}}
