\name{convert2Mcmc}
\alias{convert2Mcmc}
\title{Convert samples to mcmc objects...}
\usage{convert2Mcmc(samples, mcmcOptions)}
\description{Convert samples to mcmc objects}
\value{an S3 class \dQuote{mcmc} object}
\keyword{programming}
\arguments{\item{samples}{samples matrix or vector}
\item{mcmcOptions}{the \code{\linkS4class{McmcOptions}} object which was chosen for the
production of \code{samples}}}
