% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPrior-classes.R
\docType{class}
\name{CustomGPrior-class}
\alias{CustomGPrior-class}
\title{The custom g-prior class}
\description{
This class wraps around a custom log prior density for the covariance factor g.
}
\seealso{
the constructor \code{\link{CustomGPrior}}
}
\keyword{classes}
