\name{CustomGPrior}
\alias{CustomGPrior}
\title{Constructor for the custom g-prior class...}
\usage{CustomGPrior(logDens)}
\description{Constructor for the custom g-prior class}
\value{a new \code{\linkS4class{CustomGPrior}} object}
\keyword{classes}
\arguments{\item{logDens}{the log prior density function for g}}
