\name{Extract.GlmBayesMfp}
\title{Extract method for GlmBayesMfp objects...}
\usage{\method{[}{GlmBayesMfp} (x, ...)}
\description{Extract method for GlmBayesMfp objects}
\details{Extract a subset of models from a \code{\link{GlmBayesMfp}} object.}
\value{The subsetted object.}
\alias{Extract.GlmBayesMfp}
\alias{[.GlmBayesMfp}
\keyword{methods}
\seealso{\code{\link{glmBayesMfp}}}
\author{Daniel Sabanes Bove \email{daniel.sabanesbove@ifspm.uzh.ch}}
\arguments{\item{x}{valid \code{\link{GlmBayesMfp}} object}
\item{\dots}{transports the indexes of the models}}
