% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{bfp}
\alias{bfp}
\alias{uc}
\title{Mark a covariate for transformation with fractional polynomials}
\usage{
bfp(x, max = 2, scale = TRUE, rangeVals = NULL)

uc(x)
}
\arguments{
\item{x}{the covariate name}

\item{max}{maximum degree for this FP}

\item{scale}{use pre-transformation shifting and scaling to avoid numerical problems?}

\item{rangeVals}{extra numbers if the shifting and scaling should consider values in this range}
}
\value{
name of the provided covariate, with the other function parameters as  attached attributes

Just the name of the provided covariate
}
\description{
Using this function, you can mark a covariate for transformation with fractional polynomials
in the formula for \code{\link{glmBayesMfp}}.

Using this function, you can mark a covariate or a group of combined covariates for joint
variable selection (\dQuote{uncertain covariate fixed form covariate groups}) in the formula for
\code{\link{glmBayesMfp}}.
}
\keyword{utilities}
