% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpTrans.R
\name{fpTrans}
\alias{fpTrans}
\title{Transform formula variables}
\usage{
fpTrans(var, powers = 1, scale = 1, shift = 0, center = TRUE)
}
\arguments{
\item{var}{the variable to transform}

\item{powers}{one or more powers}

\item{scale}{value to scale the variable after shifting (default=1)}

\item{shift}{value to shift the variable by (default=0)}

\item{center}{center the variable when tranforming.}
}
\value{
the transformed vector
}
\description{
Simple function to apply the Box Tidwell transformation to a variables in a formula.
Variable is first shifted and scaled
NewVar = (Var+shift)/scale
then transformed and optionally centered.
Can be used in formulas as poly() is.
}
\keyword{utilities}
