% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncenteredDesignMatrix.R
\name{getUncenteredDesignMatrix}
\alias{getUncenteredDesignMatrix}
\title{Construct the design matrix for a given bfp GLM model}
\usage{
getUncenteredDesignMatrix(
  modelConfig = object[[1]]$configuration,
  object,
  fixedColumns = TRUE
)
}
\arguments{
\item{modelConfig}{the model configuration list which must have elements
\dQuote{powers} and \dQuote{powers}. Defaults to the configuration of the first element of}

\item{object}{the \code{GlmBayesMfp} object, which is needed because it contains
the covariates matrix and indices vector}

\item{fixedColumns}{return the fixed columns inside the matrix (default) or not?}
}
\value{
The design matrix, where the non-fixed part is columnwise centered (that is, the
colmeans are zero).
}
\description{
This is an internal function to construct the UNCENTERED design matrix for a given
bfp GLM model.
}
\keyword{internal}
\keyword{utilities}
