% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPrior-classes.R
\docType{class}
\name{GPrior-class}
\alias{GPrior-class}
\title{The virtual g-prior class}
\description{
This is the virtual g-prior class from which other g-prior classes derive.
The slots are:
\describe{
\item{logDens}{the prior log density}
}
}
\seealso{
\code{\linkS4class{HypergPrior}}, \code{\linkS4class{InvGammaGPrior}},
\code{\linkS4class{IncInvGammaGPrior}}, \code{\linkS4class{CustomGPrior}}
}
\keyword{classes}
\keyword{internal}
