% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPrior-classes.R
\docType{class}
\name{HypergPrior-class}
\alias{HypergPrior-class}
\title{The hyper-g prior class}
\description{
The slots are:
\describe{
\item{a}{the hyperparameter}
}
}
\seealso{
the constructor \code{\link{HypergPrior}}
}
\keyword{classes}
