\name{HypergPrior-initialize}
\title{Initialization method for the "HypergPrior" class...}
\usage{\S4method{initialize}{HypergPrior}(.Object, a, \dots)}
\description{Initialization method for the "HypergPrior" class}
\value{the initialized object}
\alias{HypergPrior-initialize}
\alias{initialize,HypergPrior-method}
\keyword{methods}
\keyword{internal}
\author{Daniel Sabanes Bove \email{daniel.sabanesbove@ifspm.uzh.ch}}
\arguments{\item{.Object}{the \code{\linkS4class{HypergPrior}} we want to initialize}
\item{a}{the hyperparameter value}
\item{\dots}{unused}}
