\name{IncInvGammaGPrior-class}
\alias{IncInvGammaGPrior-class}
\title{The incomplete inverse gamma g-prior class}
\description{
  The slots are: \describe{ \item{a}{the first
  hyperparameter} \item{b}{the second hyperparameter} }
}
\seealso{
  the constructor \code{\link{IncInvGammaGPrior}}
}
\keyword{classes}

