\name{IncInvGammaGPrior-initialize}
\alias{IncInvGammaGPrior-initialize}
\alias{initialize,IncInvGammaGPrior-method}
\title{Initialization method for the "IncInvGammaGPrior" class}
\arguments{
  \item{.Object}{the \code{\linkS4class{IncInvGammaGPrior}}
  we want to initialize}

  \item{a}{the first hyperparameter value}

  \item{b}{the second hyperparameter value}

  \item{\dots}{unused}
}
\value{
  the initialized object
}
\description{
  Initialization method for the "IncInvGammaGPrior" class
}
\author{
  Daniel Sabanes Bove
  \email{daniel.sabanesbove@ifspm.uzh.ch}
}
\keyword{internal}
\keyword{methods}

