% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPrior-classes.R
\docType{class}
\name{InvGammaGPrior-class}
\alias{InvGammaGPrior-class}
\title{The inverse gamma g-prior class}
\description{
The slots are:
\describe{
\item{a}{the first hyperparameter}
\item{b}{the second hyperparameter}
}
}
\seealso{
the constructor \code{\link{InvGammaGPrior}}
}
\keyword{classes}
