\name{InvGammaGPrior-initialize}
\title{Initialization method for the "InvGammaGPrior" class...}
\usage{\S4method{initialize}{InvGammaGPrior}(.Object, a, b, \dots)}
\description{Initialization method for the "InvGammaGPrior" class}
\value{the initialized object}
\alias{InvGammaGPrior-initialize}
\alias{initialize,InvGammaGPrior-method}
\keyword{methods}
\keyword{internal}
\author{Daniel Sabanes Bove \email{daniel.sabanesbove@ifspm.uzh.ch}}
\arguments{\item{.Object}{the \code{\linkS4class{InvGammaGPrior}} we want to initialize}
\item{a}{the first hyperparameter value}
\item{b}{the second hyperparameter value}
\item{\dots}{unused}}
