\name{InvGammaGPrior}
\alias{InvGammaGPrior}
\title{Constructor for the inverse gamma g-prior class...}
\usage{InvGammaGPrior(a=0.001, b=0.001)}
\description{Constructor for the inverse gamma g-prior class}
\value{a new \code{\linkS4class{InvGammaGPrior}} object}
\keyword{classes}
\arguments{\item{a}{the first positive hyperparameter (default: 0.001)}
\item{b}{the second positive hyperparameter (default: 0.001)}}
