\name{logMargLiks}
\alias{logMargLiks}
\title{Extract the log marginal likelihood estimates from a GlmBayesMfp object...}
\usage{logMargLiks(GlmBayesMfpObject)}
\description{Extract the log marginal likelihood estimates from a GlmBayesMfp object}
\value{the vector of log marginal likelihood estimates}
\keyword{utilities}
\arguments{\item{GlmBayesMfpObject}{the object}}
