\name{logPriors}
\alias{logPriors}
\title{Extract the log prior values from a GlmBayesMfp object...}
\usage{logPriors(GlmBayesMfpObject)}
\description{Extract the log prior values from a GlmBayesMfp object}
\value{the vector of log prior values}
\keyword{utilities}
\arguments{\item{GlmBayesMfpObject}{the object}}
