% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/McmcOptions-class.R
\docType{class}
\name{McmcOptions-class}
\alias{McmcOptions-class}
\title{Class for the three canonical MCMC options}
\description{
The slots are:
\describe{
\item{iterations}{number of MCMC iterations}
\item{burnin}{number of burn-in iterations which are not saved}
\item{step}{only every step-th iteration is saved after the burn-in}
}
}
\keyword{classes}
\keyword{internal}
