\name{posteriors}
\alias{posteriors}
\title{Extract posterior model probability estimates from a GlmBayesMfp object...}
\usage{posteriors(GlmBayesMfpObject, type=c("normalized", "sampling"))}
\description{Extract posterior model probability estimates from a GlmBayesMfp object}
\value{the requested probs from all models}
\keyword{utilities}
\arguments{\item{GlmBayesMfpObject}{the object}
\item{type}{type of posterior model probability estimates to be extracted from
\code{GlmBayesMfpObject}}}
