% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictCoxTBF.R
\name{predict.TBFcox.BMA}
\alias{predict.TBFcox.BMA}
\title{Prediction methods for CoxTBF objects for BMA models}
\usage{
\method{predict}{TBFcox.BMA}(object, newdata, times, ...)
}
\arguments{
\item{object}{a model fitted with \code{\link{coxTBF}}}

\item{newdata}{a dataframe with the same variables as the original data used to fit the object}

\item{times}{a vector of times to predict survival probability for}

\item{...}{not used.}
}
\value{
A data frame of survival probabilities with rows for each row of newdata and columns for each time.
}
\description{
Predicts survival probabilities at given times. Compatible with predictSurvProb functions 
required by \code{pec} package. Predicts BMA objects.
}
