\name{print.GlmBayesMfp}
\alias{print.GlmBayesMfp}
\title{Print a GlmBayesMfp object.}
\usage{\method{print}{GlmBayesMfp} (x, ...)}
\description{Print a GlmBayesMfp object.}
\value{Only used for its side effect}
\keyword{methods}
\seealso{\code{\link{glmBayesMfp}}}
\author{Daniel Sabanes Bove \email{daniel.sabanesbove@ifspm.uzh.ch}}
\arguments{\item{x}{valid \code{\link{GlmBayesMfp}} object}
\item{\dots}{unused}}
