\name{sampleSize}
\alias{sampleSize}
\title{Compute the number of samples for a given MCMC options triple...}
\usage{sampleSize(mcmcOptions)}
\description{Compute the number of samples for a given MCMC options triple}
\value{the resulting sample size}
\keyword{programming}
\arguments{\item{mcmcOptions}{the \code{\linkS4class{McmcOptions}} object}}
