\name{uc}
\alias{uc}
\title{Mark a covariate expression for joint variable selection...}
\usage{uc(x)}
\description{Mark a covariate expression for joint variable selection}
\details{Using this function, you can mark a covariate or a group of combined covariates for joint
variable selection (\dQuote{uncertain covariate fixed form covariate groups}) in the formula for
\code{\link{glmBayesMfp}}.}
\value{Just the name of the provided covariate}
\keyword{utilities}
\arguments{\item{x}{the covariate name}}
