% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeFormula.R
\name{writeFormula}
\alias{writeFormula}
\title{Construct a survival formula based on a glmBfp object with censInd not null.}
\usage{
writeFormula(models.listpart, time.var, status.var)
}
\arguments{
\item{models.listpart}{the glmBfp object for which to construct the survival formula}

\item{time.var}{the name of the time variable as character string}

\item{status.var}{the name of the censoring indicator variable as character string}
}
\value{
The formula object based on the glmBfp object.
}
\description{
This is an internal function to construct a survival formula based on a glmBfp object
with censInd not null.
}
\keyword{internal}
\keyword{utilities}
