% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFpTransforms.R
\name{boxTidwell}
\alias{boxTidwell}
\alias{\%bt\%}
\alias{'\%bt\%'}
\title{Box Tidwell transformation}
\usage{
x \%bt\% pow
}
\arguments{
\item{x}{the numeric vector}

\item{pow}{the single power}
}
\value{
the transformed vector
}
\description{
Simple function to apply the Box Tidwell transformation to a vector, when a (single) power is
provided.
}
\keyword{internal}
\keyword{utilities}
