\name{boxTidwell}
\alias{boxTidwell}
\alias{'\%bt\%'}
\title{Box Tidwell transformation}
\usage{
  %bt%(x, pow)
}
\arguments{
  \item{x}{the numeric vector}

  \item{pow}{the single power}
}
\value{
  the transformed vector
}
\description{
  Simple function to apply the Box Tidwell transformation
  to a vector, when a (single) power is provided.
}
\keyword{internal}
\keyword{utilities}

