% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/McmcOptions-methods.R
\name{convert2Mcmc}
\alias{convert2Mcmc}
\title{Convert samples to mcmc objects}
\usage{
convert2Mcmc(samples, mcmcOptions)
}
\arguments{
\item{samples}{samples matrix or vector}

\item{mcmcOptions}{the \code{\linkS4class{McmcOptions}} object which was chosen for the
production of \code{samples}}
}
\value{
an S3 class \dQuote{mcmc} object
}
\description{
Convert samples to mcmc objects
}
\keyword{programming}
