\name{CustomGPrior}
\alias{CustomGPrior}
\title{Constructor for the custom g-prior class}
\usage{
  CustomGPrior(logDens)
}
\arguments{
  \item{logDens}{the log prior density function for g}
}
\value{
  a new \code{\linkS4class{CustomGPrior}} object
}
\description{
  Constructor for the custom g-prior class
}
\keyword{classes}

