% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlmBayesMfp-methods.R
\name{Extract.GlmBayesMfp}
\alias{Extract.GlmBayesMfp}
\alias{[.GlmBayesMfp}
\title{Extract method for GlmBayesMfp objects}
\usage{
\method{[}{GlmBayesMfp}(x, ...)
}
\arguments{
\item{x}{valid \code{\link{GlmBayesMfp}} object}

\item{\dots}{transports the indexes of the models}
}
\value{
The subsetted object.
}
\description{
Extract a subset of models from a \code{\link{GlmBayesMfp}} object.
}
\seealso{
\code{\link{glmBayesMfp}}
}
\author{
Daniel Sabanes Bove \email{daniel.sabanesbove@ifspm.uzh.ch}
}
\keyword{methods}
