\name{Extract.GlmBayesMfp}
\alias{Extract.GlmBayesMfp}
\alias{"[.GlmBayesMfp"}
\title{Extract method for GlmBayesMfp objects}
\arguments{
  \item{x}{valid \code{\link{GlmBayesMfp}} object}

  \item{\dots}{transports the indexes of the models}
}
\value{
  The subsetted object.
}
\description{
  Extract a subset of models from a
  \code{\link{GlmBayesMfp}} object.
}
\author{
  Daniel Sabanes Bove
  \email{daniel.sabanesbove@ifspm.uzh.ch}
}
\seealso{
  \code{\link{glmBayesMfp}}
}
\keyword{methods}

