% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpScale.R
\name{fpScale}
\alias{fpScale}
\title{Shift and scale a covariate vector (if wished) to have positive and small numbers.}
\usage{
fpScale(x, scaling = TRUE)
}
\arguments{
\item{x}{the covariate vector}

\item{scaling}{should shifting and scaling be performed (default)?}
}
\value{
list of \sQuote{shift} and \sQuote{scale} parameters appropriate for this
covariate vector.
}
\description{
This function is (almost exactly) copied from package \dQuote{mfp} to be consistent.
}
\keyword{internal}
