% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFpTransforms.R
\name{getFpTransforms}
\alias{getFpTransforms}
\title{Get the FP transforms matrix of a given covariate vector}
\usage{
getFpTransforms(vec, powers, center = TRUE)
}
\arguments{
\item{vec}{positive (== already shifted and scaled) column vector with proper colname}

\item{powers}{power vector with at least one element}

\item{center}{center the columns of the FP transforms matrix around zero? (default)}
}
\value{
the FP transforms matrix with proper colnames.
}
\description{
Get the (centered) FP transforms matrix of a given covariate vector, when the corresponding 
power vector (with at least one element) is provided.
}
\keyword{internal}
\keyword{utilities}
