% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLogGPrior.R
\name{getLogGPrior}
\alias{getLogGPrior}
\title{Helper function for glmBayesMfp: Returns the normalized log g prior density}
\usage{
getLogGPrior(gPrior)
}
\arguments{
\item{gPrior}{the gPrior argument passed to \code{\link{glmBayesMfp}}}
}
\value{
The R function.
}
\description{
Returns the normalized log of the prior density on g as an R function.
It is checked that the result is valid, that is that exp(f) integrates to 1 over the
positive real line.
}
\keyword{internal}
