% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getModelCoefs.R
\name{getModelCoefs}
\alias{getModelCoefs}
\title{Estimate shrunken coefficients from GlmBayesMfp object for Cox model}
\usage{
getModelCoefs(model.listpart, mcmc, sep = FALSE)
}
\arguments{
\item{sep}{should coefficients be returned separately (default=FALSE)}

\item{models.listpart}{the glmBfp object for which to construct the survival formula}

\item{...}{additional arguments to pass to \code{\link{sampleBma}}}
}
\value{
A named vector of coefficients.
}
\description{
This is an internal function to estimate shrunken coefficients from GlmBayesMfp object
for Cox models. It calls \code{\link{sampleBma}} and then calculates coefficents based
on a linear fit.
}
\keyword{internal}
\keyword{utilities}
