% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlmBayesMfpSamples-methods.R
\name{GlmBayesMfpSamples-subsetting}
\alias{GlmBayesMfpSamples-subsetting}
\alias{[,GlmBayesMfpSamples,ANY,missing,missing-method}
\title{Subset method for GlmBayesMfpSamples objects}
\usage{
\S4method{[}{GlmBayesMfpSamples,ANY,missing,missing}(x, i)
}
\arguments{
\item{x}{the \code{\linkS4class{GlmBayesMfpSamples}} object}

\item{i}{the vector defining the subset of samples}
}
\value{
The subset of the same class.
}
\description{
Index the samples to select a subset of samples.
}
\note{
The function call will fail if any of the saved bfpCurves or ucCoefs
does not have enough samples to be subset by \code{i} !
}
\seealso{
\code{\linkS4class{GlmBayesMfpSamples}}
}
\keyword{methods}
