% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmBfp-package.R
\docType{package}
\name{glmBfp-package}
\alias{glmBfp-package}
\alias{glmBfp}
\title{Bayesian inference for fractional polynomial models from the GLM
and Cox family}
\description{
Bayesian inference for fractional polynomial models from the GLM and Cox
family
}
\author{
Daniel Sabanes Bove \email{daniel.sabanesbove@ifspm.uzh.ch}
}
\keyword{package}
