% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmBfp-package.R
\docType{package}
\name{glmBfp-package}
\alias{glmBfp}
\alias{glmBfp-package}
\title{Bayesian inference for fractional polynomial models from the GLM and Cox
family}
\description{
Implements the Bayesian paradigm for fractional polynomials in generalized linear models, described by Held, Sabanes Bove, Gravestock(2015) \doi{10.1214/14-STS510} and Cox models described by Held, Gravestock, Sabanes Bove (2016) \doi{10.1002/sim.7090}. See package 'bfp' for the treatment of normal models.
}
\references{
\itemize{
 \item Leonhard Held, Daniel Sabanés Bové, and Isaac Gravestock. "Approximate Bayesian model selection with the deviance statistic." Statistical Science (2015): 242-257.
 \item Leonhard Held, Isaac Gravestock, and Daniel Sabanés Bové. "Objective Bayesian model selection for Cox regression." Statistics in Medicine 35.29 (2016): 5376-5390.
}
}
\author{
\strong{Maintainer}: Isaac Gravestock \email{isaac.gravestock@gmail.com}

Authors:
\itemize{
  \item Daniel Sabanes Bove
}

Other contributors:
\itemize{
  \item David Maisonave [contributor]
  \item Terry Therneau [contributor]
  \item R Core Team [contributor]
}

}
\keyword{internal}
