\docType{package}
\name{glmBfp-package}
\alias{glmBfp}
\alias{glmBfp-package}
\title{Bayesian inference for fractional polynomial models from the GLM family}
\description{
  Bayesian inference for fractional polynomial models from
  the GLM family
}
\author{
  Daniel Sabanes Bove
  \email{daniel.sabanesbove@ifspm.uzh.ch}
}
\keyword{package}

