\name{HypergPrior-initialize}
\alias{HypergPrior-initialize}
\alias{initialize,HypergPrior-method}
\title{Initialization method for the "HypergPrior" class}
\arguments{
  \item{.Object}{the \code{\linkS4class{HypergPrior}} we
  want to initialize}

  \item{a}{the hyperparameter value}

  \item{\dots}{unused}
}
\value{
  the initialized object
}
\description{
  Initialization method for the "HypergPrior" class
}
\author{
  Daniel Sabanes Bove
  \email{daniel.sabanesbove@ifspm.uzh.ch}
}
\keyword{internal}
\keyword{methods}

