% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inclusionProbs.R
\name{inclusionProbs}
\alias{inclusionProbs}
\title{Compute posterior inclusion probabilites based on GlmBayesMfp object}
\usage{
inclusionProbs(
  GlmBayesMfpObject,
  postProbs = posteriors(GlmBayesMfpObject, type = "normalized")
)
}
\arguments{
\item{GlmBayesMfpObject}{the GlmBayesMfp object}

\item{postProbs}{the vector of posterior model probabilities, defaults to
the normalized probabilities in \code{GlmBayesMfpObject}}
}
\value{
the resulting inclusion probabilities vector
}
\description{
Compute (model averaged) posterior inclusion probabilites for the uncertain
variables (including FP variables) based on a GlmBayesMfp object.
}
\keyword{utilities}
