\name{InvGammaGPrior-initialize}
\alias{"initialize,InvGammaGPrior-method"}
\alias{"InvGammaGPrior-initialize"}
\alias{InvGammaGPrior-initialize}
\title{Initialization method for the "InvGammaGPrior" class}
\usage{
  \S4method{initialize}{InvGammaGPrior}(.Object, a, b,
  \dots)
}
\arguments{
  \item{.Object}{the \code{\linkS4class{InvGammaGPrior}} we
  want to initialize}

  \item{a}{the first hyperparameter value}

  \item{b}{the second hyperparameter value}

  \item{\dots}{unused}
}
\value{
  the initialized object
}
\description{
  Initialization method for the "InvGammaGPrior" class
}
\author{
  Daniel Sabanes Bove
  \email{daniel.sabanesbove@ifspm.uzh.ch}
}
\keyword{internal}
\keyword{methods}

