% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPrior-classes.R
\name{InvGammaGPrior}
\alias{InvGammaGPrior}
\title{Constructor for the inverse gamma g-prior class}
\usage{
InvGammaGPrior(a = 0.001, b = 0.001)
}
\arguments{
\item{a}{the first positive hyperparameter (default: 0.001)}

\item{b}{the second positive hyperparameter (default: 0.001)}
}
\value{
a new \code{\linkS4class{InvGammaGPrior}} object
}
\description{
Constructor for the inverse gamma g-prior class
}
\keyword{classes}
