% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{logMargLiks}
\alias{logMargLiks}
\title{Extract the log marginal likelihood estimates from a GlmBayesMfp object}
\usage{
logMargLiks(GlmBayesMfpObject)
}
\arguments{
\item{GlmBayesMfpObject}{the object}
}
\value{
the vector of log marginal likelihood estimates
}
\description{
Extract the log marginal likelihood estimates from a GlmBayesMfp object
}
\keyword{utilities}
