% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{logPriors}
\alias{logPriors}
\title{Extract the log prior values from a GlmBayesMfp object}
\usage{
logPriors(GlmBayesMfpObject)
}
\arguments{
\item{GlmBayesMfpObject}{the object}
}
\value{
the vector of log prior values
}
\description{
Extract the log prior values from a GlmBayesMfp object
}
\keyword{utilities}
