% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/McmcOptions-class.R
\name{McmcOptions}
\alias{McmcOptions}
\title{Constructor for class McmcOptions}
\usage{
McmcOptions(
  iterations = as.integer(burnin + (step * samples)),
  burnin = 10000L,
  step = 10L,
  samples = 10000L
)
}
\arguments{
\item{iterations}{number of MCMC iterations (default: \code{110,000})}

\item{burnin}{number of burn-in iterations which are not saved (default:
\code{10,000})}

\item{step}{only every step-th iteration is saved after the burn-in
(default: \code{10})}

\item{samples}{number of resulting samples (by default \code{10,000} will result)}
}
\value{
the freshly built object of class \code{\linkS4class{McmcOptions}}
}
\description{
Note that the argument \code{samples} is included for convenience only -
you can specify it instead of \code{iterations}.
}
\keyword{classes}
