% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{posteriors}
\alias{posteriors}
\title{Extract posterior model probability estimates from a GlmBayesMfp object}
\usage{
posteriors(GlmBayesMfpObject, type = c("normalized", "sampling"))
}
\arguments{
\item{GlmBayesMfpObject}{the object}

\item{type}{type of posterior model probability estimates to be extracted from
\code{GlmBayesMfpObject}}
}
\value{
the requested probs from all models
}
\description{
Extract posterior model probability estimates from a GlmBayesMfp object
}
\keyword{utilities}
