% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlmBayesMfp-methods.R
\name{print.GlmBayesMfp}
\alias{print.GlmBayesMfp}
\title{Print a GlmBayesMfp object.}
\usage{
\method{print}{GlmBayesMfp}(x, ...)
}
\arguments{
\item{x}{valid \code{\link{GlmBayesMfp}} object}

\item{\dots}{unused}
}
\value{
Only used for its side effect
}
\description{
Print a GlmBayesMfp object.
}
\seealso{
\code{\link{glmBayesMfp}}
}
\author{
Daniel Sabanes Bove \email{daniel.sabanesbove@ifspm.uzh.ch}
}
\keyword{methods}
